pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-generator"
version = "1.1.0"
description = "Automated PR creation from branch patterns across GitHub and Bitbucket"
requires-python = ">=3.11"
dependencies = [
    "requests==2.32.5",
    "PyJWT[crypto]==2.12.1",
    "cryptography==46.0.5",
    "pyyaml==6.0.3",
]

[project.scripts]
pr-generator = "pr_generator.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
omit = ["src/pr_generator/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
